refactor : 체크박스 컴포넌트 위치 변경
