+++ src/components/user/WithdrawalForm.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 
-import StyledCheckbox from 'components/user/StyledCheckbox';
+import StyledCheckbox from 'components/common/StyledCheckbox';
 import useSelect from 'hooks/useSelect';
 import { deleteAccount } from 'api/user';
